chore(build): add new matrix workflow
This adds a new behavior of buildcheck and release workflow.
The previous generated array for the matrix step,
will now be generated from `setup_matrix.py`, this script
reads `.github/workflow_config.yml` and generates the array
according to it's setup.

This allows an easy to maintain entrypoint to choose which
SBC will be build on test builds and releases.

It also includes version bump from `action/checkout@v3` to
`action/checkout@v4`

Updated `.gitignore` to ignore possible `.vscode` directory

Signed-off-by: Stephan Wendel <me@stephanwe.de>